pyproject.toml
[project]
name = "interopar-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "bcrypt==4.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
    "pyodbc>=5.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
